[codex-api] isolate websocket auth state handling [ci changed_files]
